+++ services/web/server/src/simcore_service_webserver/exporter/formatter/xlsx/core/styling_components.py
